Refactor the tests to increase readability
* Move utils to separate files
* Replace { waitUntilInitialized: true } in getLikely() with waitUntilInitialized()
* Shorten the expect() require
