+++ test/utils/waitUntilLikelyInitialized.js
@@ -1,0 +1,11 @@
+/* eslint-env node, mocha, es6 */
+
+'use strict';
+
+const until = require('selenium-webdriver/lib/until');
+
+function waitUntilLikelyInitialized(driver) {
+    return driver.wait(until.elementLocated({ css: '.likely_ready' }), 1500);
+}
+
+module.exports = waitUntilLikelyInitialized;
